+++ extensions/sphinx/Query.php
@@ -668,7 +668,7 @@ class Query extends Component implements QueryInterface
 	 * Fills the query result rows with the snippets built from source determined by
 	 * [[snippetCallback]] result.
 	 * @param array $rows raw query result rows.
-	 * @return array query result rows with filled up snippets.
+	 * @return array|ActiveRecord[] query result rows with filled up snippets.
 	 */
 	protected function fillUpSnippets($rows)
 	{