+++ framework/db/ActiveQueryTrait.php
@@ -132,9 +132,10 @@ trait ActiveQueryTrait
 	/**
 	 * Converts found rows into model instances
 	 * @param array $rows
+	 * @param bool $callAfterFind
 	 * @return array|ActiveRecord[]
 	 */
-	private function createModels($rows)
+	private function createModels($rows, $callAfterFind = true)
 	{
 		$models = [];
 		if ($this->asArray) {
@@ -154,11 +155,11 @@ trait ActiveQueryTrait
 			$class = $this->modelClass;
 			if ($this->indexBy === null) {
 				foreach ($rows as $row) {
-					$models[] = $class::create($row);
+					$models[] = $class::create($row, $callAfterFind);
 				}
 			} else {
 				foreach ($rows as $row) {
-					$model = $class::create($row);
+					$model = $class::create($row, $callAfterFind);
 					if (is_string($this->indexBy)) {
 						$key = $model->{$this->indexBy};
 					} else {