+++ extensions/redis/CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 redis extension Change Log
 2.0.0 beta under development
 ----------------------------
 
+- Bug #1993: afterFind event in AR is now called after relations have been populated (cebe)
 - Enh #1773: keyPrefix property of Session and Cache is not restricted to alnum characters anymore (cebe)
 
 2.0.0 alpha, December 1, 2013